requirements-models.txt
# Compatible versions for TinyIMGserver
torch>=2.0.0,<2.1.0
torchvision>=0.15.0,<0.16.0
torchaudio>=2.0.0,<2.1.0
diffusers>=0.21.0,<0.22.0
transformers>=4.25.0,<4.35.0
accelerate>=0.20.0,<0.24.0
peft>=0.6.0,<0.8.0
huggingface_hub>=0.16.0,<0.18.0
safetensors>=0.3.1
Pillow>=9.0.0
